Refactor backup_repo.sh script to improve directory creation
The backup_repo.sh script has been refactored to improve the directory creation process. The change ensures that the backup directory is created using the -p flag, which creates parent directories if they do not exist. This ensures that the script works correctly even if the backup directory does not exist.

Fixes #34